Rename and properly use function SetPeriod.
Set* is the naming scheme I'm using for easier file-system
recognition, and it makes sense to keep it as well for the actual
function name as well.

While we're at it, let's be consistent and remove the definition
from the .dat file so that the populated data is guaranteed to be a
copy.  Small Pyomo quirk that it doesn't give us a NOTICE if we have
multiple definitions of something.


git-svn-id: http://svn.temoaproject.org/temoa/trunk@13 3acdfd15-2c5e-48eb-ba72-2386a29e86c2