Update to use AMPL set slices.
Sweet!  The AMPL syntax of

set  someset[slice_a] := ...;
set  someset[slice_b] := ...;

 allows for a much cleaner model (elc.py implementation).  This lets 
us segment the technologies in a less artificial manner, and also 
reduces our definition of 3 constraints to one.  (It's still 
expanded to the same number of constraints in the end, but now it 
reads much more like the math we write on paper.)

Note that this /corrects/ our thinking of slices and segments.  We 
wanted segments, which are each a slice of tech.  Doing it this way 
lets the model automatically slice as necessary.  Cleaner.


git-svn-id: http://svn.temoaproject.org/temoa/trunk@16 3acdfd15-2c5e-48eb-ba72-2386a29e86c2