Convert param to generated set for efficiency
Moving the Investment parameter to a set definition.  This has
implications for efficiency because expression generation is the
bottleneck of Pyomo.  By generating effectively a sparse matrix, we
can do a cheap "does this exist" instead of an expensive operation,
then only to multiply it by 0.  It's the same result, just cheaper
computationally, if not less intuitive.


git-svn-id: http://svn.temoaproject.org/temoa/trunk@19 3acdfd15-2c5e-48eb-ba72-2386a29e86c2