Make objective function clearer via parameterization
By precalculating some of the model parameters with a rule, we can 
clean up the objective function to four nested loops with now just a 
few lookups.  Though it isn't at first as perhaps intuitive as the 
math on paper, this changeset makes the code much easier to read. 
(For my mistake earlier, this needs to be compared against the 
objective.py as changed between r10 and r11.)


git-svn-id: http://svn.temoaproject.org/temoa/trunk@20 3acdfd15-2c5e-48eb-ba72-2386a29e86c2