Defensive programming: use preposterous data
As we laboriously found out, when reading node-based data, PySP
begins with the leaf node, and works its way up to the parent.  The
problem is that it /also/ overwrites any previously set parameters
and does not issue a warning.  If we had used erroneous data where
we knew it should be overwritten, this would have perhaps clued us
in earlier to what was going.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@79 3acdfd15-2c5e-48eb-ba72-2386a29e86c2