Hard coded badness.  :-(
This was originally a hack while I was learning Pyomo, and for now
it stays.  However, this was the source of much frustration today,
and should be moved to a much less subtle spot.  This is not the
first time that I have "fixed" this issue, nor is it the first time
we've spent large amounts of time tracking it down.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@85 3acdfd15-2c5e-48eb-ba72-2386a29e86c2