Keep ReferenceModel.dat as a feasible input.
Since we often use Pyomo directly to test our model, it's best to
keep ReferenceModel.dat as a feasible input.  However, we have to
remember (or I need to update the generated_scenario_tree.py script)
to take out values from R.dat (currently a copy of
ReferenceModel.dat) so that they do not conflict with the leaf
nodes.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@86 3acdfd15-2c5e-48eb-ba72-2386a29e86c2