Update some planned model logic.
This will likely change as we work through the actual implementing,
but for now, this includes the notion of a longer history and future
than about which we actually want to use for results interpretation.

Specifically:
* the 'window_period' parameter, used to initialize three ...
* 3 internal Sets: 'exist_period', 'future_period', and
    'beyond_horizon_period'
* also adds to more derived sets, 'optimize_period', and
  'exist_vintage', which is the same as exist_period, but used for
  easier bookkeeping/code reading

In discussion of this, also realized that the expected
implementation of the activity and capacity variables is wrong.
Will address in future commits, but have sussed out the call
signature here.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@150 3acdfd15-2c5e-48eb-ba72-2386a29e86c2