Do not assume that script knows how's it being used.
Previously, a failed coopr.pyomo import assumed that nothing more could be
done.  This may not be correct, as Sphinx highlighted: the correct answer is to
give all available information, namely that there was an issue importing.  Let
the higher-up authority decide what to do.  Usually this will result in the
same behavior as before, where the system exits.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@185 3acdfd15-2c5e-48eb-ba72-2386a29e86c2