Update and add a unit correction after meeting with Joe.
* Capacity of a process is in power (energy/time)
* Activity of a process is in energy

Therefore a direct comparison is incorrect.  Combining with SegFrac
to resolve this units issue.

Also, modelers often work with end-use units as well as capacity,
which are often an order of magnitude different numerically
speaking.  Before this commit, modelers were forced to maintain a
consistent unit set, like a Capacity in GW with an Activity of 8000
GWh, instead of an Activity of 8 TWh.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@202 3acdfd15-2c5e-48eb-ba72-2386a29e86c2