Move small portion of error check function back to CommodityBalanceConstraint.
The if-statement was originally an all-in-one ordeal and I moved it
whole-sale to the temoa_lib file.  However, the if was not entirely
devoted to error checking, and the part I moved back was actually
oriented toward getting rid of a useless constraint.

Leaving the entire if in the ErrorCheck function was
once-upon-a-time fine, but Pyomo was recently updated to be more
strict (correctly so) about return values from constraint values,
and now the re-moved if is also important to correct behavior.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@212 3acdfd15-2c5e-48eb-ba72-2386a29e86c2