Add some timing results to the direct invocation method.
Pyomo's new output redirects the YAML result to a file and instead
shows meta information like timing.  That's a good precedent.

This does similar now, but instead distinguishes between stdout (for
the result) and stderr for the informational messages.

Now one can see where the model is with a command line like this:

$ ./model.py test.dat > test.sol


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@220 3acdfd15-2c5e-48eb-ba72-2386a29e86c2