A quick and dirty script to convert Pyomo-YAML to CSV
Rather than explain what it does, I'll copy here the Q&D usage
message:

The name of this script is originally rather ... verbose.  That
should reinforce that this is a "quick and dirty" script with a very
specific purpose: to convert Pyomo-based YAML results into a CSV for
easier working within a spreadsheet (or other) program.  It should
be a bit more robust than my previous attempts, but there are a
number of tidbits yet to be implemented.

To execute it, pipe YAML output into the stdin.  Any of these will
work:

$ pyomo_yaml_results_to_csv.py < results.yml
$ pyomo_yaml_results_to_csv.py < results.yml | less
$ cat results.yml | pyomo_yaml_results_to_csv.py

To put the output to a file, invoke like either of these:

$ pyomo_yaml_results_to_csv.py < results.yml > results.csv
$ cat results.yml | pyomo_yaml_results_to_csv.py > results.csv


git-svn-id: http://svn.temoaproject.org/temoa/branches/misc_scripts@228 3acdfd15-2c5e-48eb-ba72-2386a29e86c2