Be a little more robust in things non-Temoa.
I have minor hopes that this function will be useful outside of just
this project, so make it more able to cope with things non-Temoa,
like an arbitrarily named objective function.

Also update assumption of 1e-16 to 1e-15 as equivalent to 0.  1e-15
is frickin' close enough, and I've now seen a couple of 1e-16 values
fly by in the output.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@229 3acdfd15-2c5e-48eb-ba72-2386a29e86c2