No need to sort set elements.
The time_* sets are ordered, so there is no need to sort them.  This
is, in some sense, trust that Pyomo will always be robust on this
particular feature.  The risk is one I'm willing to take, because
semantically, we should either use the time_* sets as ordered, or we
shouldn't.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@235 3acdfd15-2c5e-48eb-ba72-2386a29e86c2