Implement fractional lifetime accounting.
Man, this took awhile to nail down, and there are some subtle
aspects to it, but I think it's mostly correct now.

There's one known bug, however, with the imp_* technologies: as
currently coded, their inputs are the same as their outputs, so we
end up with LP artifacts like:

 - 0.4 V_FlowIn(2010,summer,day,coal,imp_coal,2000,coal)
 - 0.4 V_FlowIn(2010,summer,day,coal,imp_coal,2010,coal)
 - 0.4 V_FlowIn(2010,summer,day,coal,t_coal,2000,electricity)
 - 0.4 V_FlowIn(2010,summer,day,coal,t_coal,2010,electricity)
 - 0.4 V_FlowIn(2010,summer,night,coal,imp_coal,2000,coal)
 - 0.4 V_FlowIn(2010,summer,night,coal,imp_coal,2010,coal)
 - 0.4 V_FlowIn(2010,summer,night,coal,t_coal,2000,electricity)
 - 0.4 V_FlowIn(2010,summer,night,coal,t_coal,2010,electricity)
 + V_FlowOut(2010,summer,day,coal,imp_coal,2000,coal)
 + V_FlowOut(2010,summer,night,coal,imp_coal,2000,coal) <= 0

This is not correct as it includes "itself" in both its inputs and
outputs.  I think the right answer is to include a "magical" ether
from which all tech_resource technologies can pull.  This would also
have the side effect of not making a generated graph with the
tech_resource technologies having an arrow to and from the resource
they provide.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@245 3acdfd15-2c5e-48eb-ba72-2386a29e86c2