Create a third commodity set: commodity_carrier.
It's poorly named but I'm having a difficult time coming up with the
correct name.

What's nice about implementing this: reduces the number of variables
over which the various variables are created, reducing solve time by
roughly 10 seconds on my machine.  The /solve/ time is the same,
because it ends up being the same number LP variables, but Pyomo
creates less internally.  Win.  Not big win, but win.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@257 3acdfd15-2c5e-48eb-ba72-2386a29e86c2