Temoa now requires at least Python v2.7.
When running via the direct invocation, we now take advantage of
several Python 2.7 features.

For now -- but this won't hold true for much longer, is my guess --
invocations with Pyomo will work with Python 2.6.

I'm not terribly worried about this requirement as Python is an
excellently engineered piece of software.  Though it has it's warts,
the more I use it, I am impressed at how well the Python community
has thought various issues through.  In that vein, installation is a
just shy of a walk in the park, for any platform.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@260 3acdfd15-2c5e-48eb-ba72-2386a29e86c2