Work around very annoying bug within Pyomo.
The bug is that when making a copy of a Set -- and by looking at the
Pyomo code, I'd believe other objects, like Params, etc. as well --
one alters the original.  Example:

M.A = Set()
M.B = A

B should be a copy of A, perhaps with the same items so as to save
memory.  Instead, that operation alters A such that any reference to
A is treated as if it B.  It's not a problem until you try to change
or work with specifically A.  Then you can't.

Workaround: make a copy of the Set yourself, rather than trust
Pyomo.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@262 3acdfd15-2c5e-48eb-ba72-2386a29e86c2