Two additional reporting variables:
* V_ActivityByPeriodInputAndTech
* V_ActivityByPeriodInputTechAndVintage

I'm trying to stick to a strict naming convention everywhere in the
code such that whenever you see an input and tech, the input always
goes on the left.  Similarly, whenever you see a tech and an output,
the output goes on the right.  So you can use a mental mnemonic
like:

inp -> tech -> out

since, in the English and math worlds, we often "think" in "left to
right" terms.

Within that, a tech may have a vintage, which collected together
make a process.  The mnemonic still holds:

inp -> (tech, vintage) -> out
inp -> process -> out

In that vein, these variable names, even though they move the
carrier (Input) in the name to the other side of tech from the
similar output variables, hold with the mnemonic.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@263 3acdfd15-2c5e-48eb-ba72-2386a29e86c2