Reorder _rule arguments, putting the model first.
This is actually a positive move by Coopr 3.0, but annoying to have
to handle.  This makes much more sense in the CS sense to have the
mandatory and guaranteed argument come first, and ought to offer a
minor speed improvement behind Coopr's scenes.  At least if coded as
I think it's coded.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@266 3acdfd15-2c5e-48eb-ba72-2386a29e86c2