Replace "return None" with "return Constraint.Skip"
This is another good, but unfortunately not backward compatible
change to work with Coopr 3.0.  Good because now functions
explicitly say what they're doing, rather than this "weird return
None" construct.

In fact, this is a form of bug prevention because if a function
doesn't return anything (on accident, perhaps), Coopr 3.0 will flag
this, as opposed to silently ignoring it.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@267 3acdfd15-2c5e-48eb-ba72-2386a29e86c2