Move initialization hack to main model file.
It's still the same hack because Pyomo doesn't offer an "official"
method of global validation and preprocessing like we need to do,
but this moves it to the main model file, and out of the Objective
function.  In this sense, it's a little "in your face" to the
modeler who may be looking at the code base.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@268 3acdfd15-2c5e-48eb-ba72-2386a29e86c2