Add a stochastically-prepared version of Temoa.
The whole point is to have the linear version of Temoa and the
stochastic version be as closely aligned as possible.  This already
possible at the expensive of a slightly less clear objective
function, which this file roughly shows.  The naming scheme is my
first, perhaps half-baked attempt to make this process generic.

In a tangential manner, this shows how one might override the
default Temoa objective function for one of their own, with perhaps
minimal fuss.  Almost as drop in as it can get.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@299 3acdfd15-2c5e-48eb-ba72-2386a29e86c2