Fix -- workaround -- to work with PySP's more picky nature.
It seems that PySP has gotten more picky since last we used it
heavily: unused variables need to be marked as such or removed
entirely from the model.  I would *love* to remove the unused
variables, but at present, I don't know how.  So deactivating them
is the best that I can do for now.

FYI: Unused variables come from the fact that we define variables
over multiple sets, but any particular combination of the sets might
not actually be used.  For instance, here's an example V_FlowIn that
might be 'legit', but would never be used:

V_FlowIn[2010, winter, day, coal, t_coal, 2030, electricity]

because we can't build 2030 vintage plants in 2010.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process@306 3acdfd15-2c5e-48eb-ba72-2386a29e86c2