An absolute (and temporary!) hack.
I need to find a more generic way to only use model variables that
exist.  Issue:

As explained in the comment, this script was written before we used
sparse sets to create variables.  That means it was telling PySP to
use variables that were previously never used, and are now not even
created.  This temporary hack uses the model instance's
understanding of what variables exist to allow or disallow it, but
is tied to a *very* specific format of this model.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@312 3acdfd15-2c5e-48eb-ba72-2386a29e86c2