Use sparse-type definition of Efficiency parameter.
A little smarter rearrangement in temoa_model.py was in order to use
a sparse-type definition of the Efficiency parameter.  At the same
time, use a sparse definition of ExistingCapacity.  Specifically, we
can initialize the global variables much earlier, which makes the
helper functions useful at an earlier point than before.

As a consequence of the sparse definition, must now employ more of
the "does key exist" guards in various places.

Also include a notice to modelers if they define an Efficiency item
of 0.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@355 3acdfd15-2c5e-48eb-ba72-2386a29e86c2