Update two rule-based parameters to also be semi-sparse.
We can deduce what are the possible indices of the DiscountRate and
LoanAnnualize parameters based on what's possible for the model to
build.  More so than saving memory or being speedier, this turns out
to be a big boon in terms of catching both useless dat file inputs
and data entry errors that Pyomo /should have/ caught.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-coopr-3.0@357 3acdfd15-2c5e-48eb-ba72-2386a29e86c2