Cleanup two functions
These two functions originally made use of the ConcreteModel to
perform there work.  This did not jive with the rest of the
ProcessesBy* functions, and made these two take an extra argument
for the model.  These functions now do their work in a single
statement with list comprehension, and, from a code perspective,
they're clearer to read.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@394 3acdfd15-2c5e-48eb-ba72-2386a29e86c2