Add stub for "DetailedModelGraph"
I'm not actually sure what this set of code from the summer Graphviz
work is doing, so until I spec it further, I'm going to leave just a
stub here.  Everytime the graph routines are run, it will display a
"Not yet implemented" message to the console to remind us (me) to
make a decision.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@406 3acdfd15-2c5e-48eb-ba72-2386a29e86c2