Add multiprocessing to Graphviz routines
With the functionalization of many of the Graphviz creation
routines, adding multiprocessing was fairly simple.  The Graphviz
generation routines will now take advantage of all cores available
to a machine.  On the very simple and small Utopia and IEW models
running on my dual-core hardware, this results in a roughly 25%
speedup of overall runtime.

With this commit, the excellent summer Graphviz work is now imported
into the Coopr3 line, with one remaining code section to review and
import.  Overall, I've only made a few minor changes to the logic,
though the code may look a little different.  Summarized below:

* Split up various graph generation routines into functions
* Dot language output (Graphviz language) "pretty-printed" for
  easier human reading, debugging, and learning.
* More efficient output through use of Dot language subgraphs
* Re-enabling of robust/flexible command line interface
* Replaced os.system() calls with subprocess.call() calls
  (portability)
* Fixed some off-by-one bugs
* Corrected a few (minor) graph connections and assumptions
* Renamed various output files
* Completed a couple of graphs
* Added multiprocessing to graphic generation.

After some testing in this branch, these functions will hit the
mainline trunk branch for all to use.  Good work, folks!


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@409 3acdfd15-2c5e-48eb-ba72-2386a29e86c2