Implement newly found DemandActivity need
The inline comment fairly well explains the use case for this
constraint:

For end-use demands, it is unreasonable to let the optimizer only
allow use in a single time slice.  For instance, if household A buys
a natural gas furnace while household B buys an electric furnace,
then both units should be used through the year.  Without this
constraint, the model might choose to only use the electric during
the day, and the natural gas during the night.

Mathematically, this constraint ensures that the ratio of the
Activity to demand is constant for all time slices.  The
multiplication trick here is analogous to what is performed in the
Baseload constraint.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@423 3acdfd15-2c5e-48eb-ba72-2386a29e86c2