Partially deal with silly user input.
Previously, the user was allowed to input their time_* set elements
in any order, but Temoa relied on an implicit assumption that they
were sorted.  This commit ensures that PeriodLength is calculated
properly, but I still need to find a way to ensure that the set
time_optimize is properly ordered.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@424 3acdfd15-2c5e-48eb-ba72-2386a29e86c2