Ensure set ordering.
The most ideal fix would be to enforce a sorting through Pyomo APIs,
but I don't know that Pyomo can currently do that.  Temoa is lucky,
however, in that we manufacture the potential problem sets from
other the sets the user supplies.  Thus, we have a shim where we can
enforce sorted ordering.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@425 3acdfd15-2c5e-48eb-ba72-2386a29e86c2