The MARKAL specifics.
The gist, which took *so* long to figure out, is the difference in
capacity calculation, for electric sector and non-electric sector
technologies.  Thus, this "general" solution to match MARKAL is to
create an electric subset of technologies that the user may specify.
With these additions, we appear, at least with the just committed
utopia-markal.dat, to agree with MARKAL to within 0.7%,


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-match-MARKAL@430 3acdfd15-2c5e-48eb-ba72-2386a29e86c2