Removing some now obsolete data files.
README explains the new Temoa stochastics interaction.  With the
update script, there is no longer a need for a general holding place
for scenario data as each generated scenario tree is put into it's
own folder, and there's no longer a need to have a specific SP
version of the model lying around.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@440 3acdfd15-2c5e-48eb-ba72-2386a29e86c2