Do not ignore non-conditional nodes
The original assumption of this script was that all nodes but the
root node were conditional.  However, the previous iteration forgot
about that assumption and did not correctly tell PySP that *all*
non-conditional nodes must be the same the across scenarios.
Instead only the first mentioned stochastic point was made to be the
same.  Now, one may specify in the options.py file the stochastic
points of a model, and the script does not force use of all
stochastic points.

It does however, correctly account for them.  When a stochastic
point is not mentioned in the stochastic_points options.py tuple, it
is also not assigned any modified data.  Therefore. the original dot
dat file is the authority on that nodes parameter values.


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@445 3acdfd15-2c5e-48eb-ba72-2386a29e86c2