Allow our Graphviz code to work in Windows.
Windows appears to have some trouble with multiple processes.  I am
not strong in Windows, and will not waste my time optimizing for it.
This is a "get it working" solution, that will also be robust.  It
will also make graph generation take noticeably longer on Windows
than it does in a Unix environment.


git-svn-id: http://svn.temoaproject.org/temoa/branches/energysystem-process-Coopr3@458 3acdfd15-2c5e-48eb-ba72-2386a29e86c2