This is the non-homogeneous Markov chain tree generator.
It needs a slightly differently formatted options.py file as well, but at this
point successfully reduces the number of scenarios from a naive power
relationship (e.g. 8^3) to a more manageable number of scenarios, given the
number of time-stages.  Also, I believe the technical term is a
"time-homogeneous Markov chain of memory 1".


git-svn-id: http://svn.temoaproject.org/temoa/branches/stochastic@466 3acdfd15-2c5e-48eb-ba72-2386a29e86c2