First cut of installer.
I don't yet know if this will go beyond Windows XP, but this is a first
cut of it.  Use it thus:

Step 1: You must manually place three files in the same directory,
glpk_*dll, glpsol.exe, and temoa.py

Step 2: Run the script, and it will create a second script with those
three files embedded.

Step 3: Run the output script on a Windows XP machine to install temoa.