An installer for Windows, and instructions.
We now have an installer for Windows, so we can offer a much better
installation experience for Windows.  Linux has always been easy.
Unfortunately, however, the Mac operating system remains a problem, and
we currently put the onus on the user to figure out how to install the
various modules on their system.