Remove mistakenly double-counted partial marginal costs.
Because of the assumption that each year within a period is homogenous
as regards the amount of energy *used*, there is no need to special
count the energy passing through dying technologies.  Doing this in the
objective function would double-count any costs for this energy.

The "nice" thing for the Utopia analysis is that there is no marginal
cost for the partial period techs (TX*), so it hasn't been an issue
during our calibration exercises.