Rethinking of how to calculate partial period techs
The overall change is to the assumption that partial-period
processes (those that have an end-of-life mid-period, rather than on a
period boundary) operate strictly until the end of their life.  Instead,
we now conform to the larger assumption that the energy demands and
configuration of use is the same for all years in a period.  The
simplification is that partial-period processes now only add a
fractional amount of their capacity to the total capacity available.
The fraction added is their operational years in the period to the whole
period.

For example, if a period is 8 years long, and a process ceases operation
3 years in, then only 3/8ths of that processes capacity are added to the
global amount available for that period.

temoa_lib:
  The only change is the generation of the sparse indices over which
  DemandConstraintIndices is called.  Specifically, replace the value
  demand with the name of the demand in the timeslice.  In practice this
  won't reduce the number of sparse indices, but uses the /name/ of the
  demand now, rather the demand value, and reduces the required
  dimensionality of the indices from 9 to 8.

temoa_model:
  Update DemandActivityConstraintIndices per temoa_lib.

  Also remove the now obsolete MARKAL-specific changes to the Temoa
  model.  Through FractionalLifeActivityLimit, Temoa's assumptions now
  agree with MARKAL on how to calculate the partial period nature of a
  processes.

temoa_rules:
  DemandActivityConstraint_rule now follows the CapacityByOutput
  modeling change.  This means that rather than using V_Activity to
  calculate end-use processes output ratios, Temoa uses the sum over
  V_FlowOut.  As this constraint is now /per demand/, it should
  correctly handle cases where a process can meet more than one end-use
  demand.

  CapacityAvailableByPeriodAndTech now accounts for the new assumption
  for partial-period processes.