Refactor to use short indices.
Prior to this commit, we were using long "descriptive" names for the
indices.  However, we have since realized that there are really on seven
indices, and believe that it is actually easier to read the code if we
follow a convention and expectation that the code is meant to be read in
tandem with our documentation.  Thus, we now use names like 'p' instead
of A_period, and it is to be understood by context to where it applies.
Generally, and unqualified letter is part of the "foreach" of the
constraint, while 'S_p' represents "Sum over the periods".

p = member of Period, or time_optimize
s = member of time_season
d = member of time_of_day
i = member of commodity or resource
t = member of all_tech
v = member of vintages
o = member of commodity or end-use demand

dem = member of demand