Very minor speed improvement.
Because some numbers may not be known at runtime, Pyomo automatically
"upgrades" numbers to objects unless specifically told otherwise.  In
the case of parameters, however, the numbers are known, so this by using
.value (or value()), we can let Python do the number crunching once, and
elide the need for Pyomo to create yet more objects.

The choice of .value over value() is because function calls are somewhat
expensive in Python speak.  However, there is one potential "gotcha":
grouping is necessary, or else the Python parser will see a
multiplication of an object by a number and upgrade the operation to an
object.  Thus a simple multiplication like Var * Param1 * Param2 could
parse like this:

x = Var * Param1.value * Param2.value
  1: Var * val1 * val2
  2: Mult(Var, val1) * val2
  3: Mult(Mult(Var, val1), val2)
  4: Mult(Var, val3)

Where val3 is the result of val1*val2.  e.g. if val1=3.3 and val2=4.7,
then val3=15.51.

The latter is what Pyomo needs to get to, and we can avoid steps 2 and 3
by grouping:

x = Var * (Param1.value * Param2.value)
  1: Var * (val1 * val2)
  2: Mult(Var, val3)))

When the Python parser sees the bare numbers in a grouped isolation, it
knows to just go ahead an multiply them together.

This is a very minor optimization, and one generally should not try to
avoid function calls except in limited cases.  I believe this to be such
an exception as it's a minor -- but noticeable -- speedup for a small
problem.  I expect this will account for about 1% to 5% speed increase
in larger models.