Implement ModelLoanLife and companion
For readability, it doesn't make sense to have two "halves" for some
parts of the objective function.  By creating a ModelLoanLife (MLL) and
TechPeriodLife (MLL and TPL in the documentation), these can be reduced
to a much more manageable one-part section.

TPL is the total length of time in a period that a process is alive.
Most items in this param will have lengths equal to the PeriodLength for
that period.

MLL is the minimum of the life of the loan for a process, or the number
of years between the vintage and end of time_optimize.