Update CapacityFactor with time slice indices.
With the realization that the capacity factor is not the same in all
time slices, we need to index it accordingly.  For example, the average
capacity factor for solar might be 48%, but it's much higher during the
daytime than it is at night.

What's not cool about this is the much more cumbersome nature of
specifying the CapacityFactor parameter.  We'll need to think at some
point about how best to have the flexibility without the headache.