Build files in temporary location.
By using /tmp, this pedagogically underscores the fact that the PDF and
HTML output are 100% buildable, and need not be saved (except for
posting for others to read, as for example, a snapshot PDF for web
downloading).

As a bonus, many machines have a RAM-backed /tmp, making builds on them
potentially much faster than using a disk-backed storage medium.