More robust pretty format code.
The previous iteration of the pformat code relied on the textual output.
For the simplistic parsing it was doing, this was slightly munging
variable and constraint names.  It was not egregious, but it was also
not correct.  This new approach uses the Coopr API to identify variables
and constraints programmaticaly, producing a more correct and consistent
output.