Provide situation-useful default solver.
We already scan for what solvers Coopr sees, so go ahead and make use of
that information to choose the default solver.  The assumption is that
if CPlex is available, it's the best.  Then Gurobi.  Then Coin Branch
and Cut (CBC), and finally defaulting to GLPK if none of those exist.