Introduce TemoaError exceptions
For certain classes of errors, we can be much nicer with the output than
the ugly stack trace that an Exception normally raises.  These are
generally known validation or sanity check errors.

In the case that someone would prefer to instrospect an error with the full stack
trace, I have not yet implemented a command line argument.  However,
they can run the model file directly via something like:

$ coopr_python  temoa_model/temoa_model.py  path/to/dat/file.dat

As opposed to preferred method of:

$ coopr_python  temoa_model/  path/to/dat/file.dat