Implement Demand distributions.
Rather than forcing the user to manually (or "off-line") calculate the
amount of demand that each period and time-slice requires, this change
allows a simple description of the distribution of an end-use demand
(EUD), and the total amount of EUD per period.

This commit adds these two parameters:

 DemandDefaultDistribution (DDD)
 DemandSpecificDistribution (DSD)

Parameter DDD is indexed by time slice (season, time of day), and
specifies the default distribution of all EUDs.  It is optional.  If a
user does not specify it, then it defaults per time slice to the
same values as SegFrac.

DDD is also partially specifiable.  That is, if a modeler would only like
to specify a subset of the time slices for EUD distribution, then so
long as the final accounting of DDD sums to 1, this is legal:

param  SegFrac  :=
 winter  day    0.2
 winter  night  0.3
 summer  day    0.3
 summer  night  0.2
	;

param  DemandDefaultDistribution  :=
 winter  day    0.4
 winter  night  0.1
	;

Temoa will automatically fill in the missing values for (summer, day)
and (summer, night) with 0.3 and 0.2, and the sum will be 1 (100% of the
time of a year).

For EUDs that do not follow the standard distribution, there is
DemandSpecificDistribution.  DSD has a third index of commodity_demand.
Like DDD, DSD is optional.  Unlike DDD, however, DSD does not allow
partial specification.  If a modeler specifies DSD for any time slice,
then the modeler must specify the DSD for all time slices.  Note that
this does mean that an EUD must be used in all time slices, just that
the modeler must ensure that the sum of all fractions of a year
specified in DSD for a particular EUD sums to 1.  For example, using the
above definition of 4 time slices (winter/summer, day/night), a modeler
might specify that heating demand is distributed only in the winter
with:

param  DemandSpecificDistribution  :=
 winter  day    heating  0.7
 winter  night  heating  0.3
	;

However, the following is not valid (because the modeler has not
specified 100% allocation of distribution), and Temoa will complain:

param  DemandSpecificDistribution  :=
 summer  day  cooling  0.6
	;

Note that for ease of typing, and respecting the 80-column width, the
code sometimes references these as DDD and DSD with

 DDD = M.DemandDefaultDistribution
 DSD = M.DemandSpecificDistribution

 # now DDD and DSD can be used in place of the longer names