Add --ecgw argument, and reorganize temoa_solve
This argument conflicts with supplying a data file because it is meant
for a evaluating a stochastic solution.  So this commit also introduces
an explanation to the user of what's going on.

Additionally, prepare temoa_solve for the two modes of operation.  This
is nice because it keeps the functionality tied to temoa_model.py.
Perhaps I can also bring the PySP functionality into the fold ...

I'm also playing with some command line colors to really bring out what
the issue is, rather than have "yet more drab gray text" to scare the
user.  Unfortunately, this is not universal: this only works on
non-Windows systems.  Fixing this so that we can get color on Windows is
on my TODO (wish)list.