This check is no longer necessary.
The original purpose for this check was because there was a bug in
Coopr.  I submitted a patch to fix that bug in Dec, 2011, but I was
hours too late as they had already started their build of Coopr.  A
year later shows the next version of Coopr released, so this is no
longer necessary.