Implement "Expected Cost of Guessing" algorithm
This is a first implementation of the ECGW algorithm.  As such, it is
has only basic functionality:

 * For each scenario, solve it, then
   * set the stage 1 variables, and solve for every other scenario.

One uses it through the --ecgw argument to the Temoa script:

    $ coopr_python  temoa_model  --help
     [...]

    $ coopr_python  temoa_model  --ecgw  path/to/stochastic/tree

It uses the PySP stochastic scenario definition files, so one only need
point it to the same files used for the stochastic analysis.

This returns a CSV file of the data, such that for each "Gambler's
Guess", it shows the total costs involved with guessing the correct
scenario, and incorrectly for each scenario, as well as the variable
values associated with each outcome.

This means that each "incorrect" scenario represents the minimum costs
involved in course correction for that scenario in question.  There is
nothing to say that the costs will be even higher, but these represent
the minimum.

The current difficult part is that Coopr is not great about leakage
between models with different data, so I have to create the temoa model
afresh for each slightly different data set.  This is a little slower
than ideal, but it works.

At this point in time, I am aware of at least one unimplemented
"biggie": parallelism.  The script is entirely serial.  The next
iteration will likely use "easy" parallelism in the form of fork()ed
processes (i.e., only up to the number of processes on the single
compute node running this script).  The bigger gains in parallelism will
come when I implement a cluster-wide ability to parallelize, but that is
not today.