Rework pformat_results for speed.
With larger models, pformat results was feeling some inefficiencies.
This implementation no longer goes back to the Pyomo model, but instead
directly uses the returned results.  For the time being, since we only
care about the variable values, this is appropriate.  At the point we
need more information reported, I'll update this code.

Meanwhile, it's pertinent to note that this code is technically not
robust, given the line that removes single quotes:

 ... name.replace("'", '') ...

The implicit assumption is that variable indices will not have single
quotes in them.  I believe this to be safe because the Pyomo AMPL data
file parser does not accept them, nor do I know if the AMPL data format
even allows them.  However, from a pure Python perspective, a single
quote is a perfectly acceptable as an index.  Thus, this is safe within
the context of Coopr, but not in a general computing sense.