Use BuildAction in place of "empty-Set" hacks.
When I first implemented these sanity checks, I was unaware of Pyomo's
BuildAction capability.  In looking at the commit logs, it appears Pyomo
/did/ have them, so shame on me.

Meanwhile, this amounts to an NFC in that the sole difference is the
lack of storage of an empty set or tuple per check.  Oh boy.  But this
is semantically better so I use it.