Work around solvers "close enough to zero answer
Solvers will terminate with a "good enough" answer.  This means that if,
to within some provable epsilon, the variable values won't change the
optimal value, then it's good enough to quit.

Meanwhile, this means we can get results in non-negative numbers that
are /almost/ zero, but techinically negative.  On the other side, we can
similarly get results that are positive and /almost/ zero.  Because we
know that our variables are positive, and we don't care about variables
less than some epsilon (1e-9 for now), just ignore these values or set
them completely to zero.